requirements.txt
python-dotenv==1.0.0
slack-sdk==3.26.2
certifi==2023.11.17